requirements.txt
llvmlite==0.21.0
tensorpack==0.8.5
matplotlib==1.4.0
numpy==1.13.3
scipy==1.0.0
seaborn==0.8.1
Pillow==5.0.0
cvxpy==0.4.11
numba==0.36.2
scikit-learn==0.19.1
PyWavelets==0.5.2
tensorflow-gpu==1.4.0
